% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_boundary.R
\name{build_boundary_CuR}
\alias{build_boundary_CuR}
\title{Title}
\usage{
build_boundary_CuR(complex)
}
\arguments{
\item{complex}{filtration complex e.g. (flt$cmplx) where flt=TDA::alphaComplexFiltration(X)}
}
\value{
BB
}
\description{
Title
}
