% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexToBoundaryF.R
\name{complex_to_boundaryF}
\alias{complex_to_boundaryF}
\title{Title}
\usage{
complex_to_boundaryF(cmplx)
}
\arguments{
\item{cmplx}{Creates boundary matrix of the filtration}
}
\value{
list of boundary matricies with out[\link{n}] being boundary matrix of n+1 simplicies
}
\description{
Title
}
