% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeRahmMap.R
\name{deRahmMap1f}
\alias{deRahmMap1f}
\title{Do DeRahm map from 0-simplicies to 1-simplicies}
\usage{
deRahmMap1f(B, X, V)
}
\arguments{
\item{B}{Boundary matrix of 1-simplicies}

\item{X}{Coordinates of the points}

\item{V}{Vectors associated with the each point}
}
\value{
DeRahm map
}
\description{
Do DeRahm map from 0-simplicies to 1-simplicies
}
