% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgeDecomp.R
\name{HodgeDecomp}
\alias{HodgeDecomp}
\title{Calculation of Hodge Decomposition of a vector field on simplicial complex}
\usage{
HodgeDecomp(BB, derahm, alg = "minres")
}
\arguments{
\item{BB}{List of boundary matricies (return of complex_to_boundaryF(...))}

\item{derahm}{DeRahm map on 1-simplicies}

\item{alg}{Algorithm for lsqr possibilities: "pinv", "minres" and "bicg"}
}
\value{
Returns a list where 1st entry is gradient field, 2nd is curl field and 3rd is harmonic the component
}
\description{
Calculation of Hodge Decomposition of a vector field on simplicial complex
}
