% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{HodgePaths}
\alias{HodgePaths}
\title{HodgePaths}
\description{
Does Hodge decomposition of descrete vector field associated with point cloud in n-dimensional space
}
\author{
Jan Stuchly & Tomas Jelinek
}
