% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{makedf}
\alias{makedf}
\title{Make dataframe from Whintney interpolation}
\usage{
makedf(out)
}
\arguments{
\item{out}{Output from fucntion WhitneyMap1f_BC}
}
\value{
returns dataframe that one can put into plotting function vplot
}
\description{
Make dataframe from Whintney interpolation
}
