% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{vplot}
\alias{vplot}
\title{Plot vector field}
\usage{
vplot(df, seg, normalize = FALSE, fx = mean)
}
\arguments{
\item{df}{dataframe with $x $y $dx, $dy being a vector field}

\item{seg}{#into how many segments do you want to average out a vector field}

\item{normalize}{TRUE or FALSE if you wish to normalize vectors when they are displayed}
}
\description{
Plot vector field
}
