% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{vplot2}
\alias{vplot2}
\title{Title}
\usage{
vplot2(df, seg, normalize = FALSE, fc = "mean", jitter = 130, n = 70)
}
\arguments{
\item{df}{dataframe with $x $y $dx, $dy being a vector field}

\item{seg}{#into how many segments do you want to average out a vector field}

\item{normalize}{TRUE or FALSE if you wish to normalize vectors when they are displayed}

\item{fc}{function for averiging vectors in cell ("mean"/"median")}

\item{jitter}{amount of jitter}

\item{n}{amount of lines}
}
\description{
Title
}
