% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhitneyMap.R
\name{WhitneyMap1f_BC}
\alias{WhitneyMap1f_BC}
\title{Title}
\usage{
WhitneyMap1f_BC(cmplx, flt, X, Rahm)
}
\arguments{
\item{cmplx}{simplicial complex}

\item{flt}{filtration complex}

\item{X}{Coordinates of points}

\item{Rahm}{Scalar mapping to 1-simplicies}
}
\value{
Returns oneforms and points of their origin
}
\description{
Title
}
